+++ main.go
@@ -82,7 +82,7 @@ func main() {
 		if *passphraseFlag == "" {
 			log.Fatal("--passphrase or $TLSPROXY_PASSPHRASE must be set")
 		}
-		if !cfg.AcceptTOS {
+		if cfg.AcceptTOS == nil || !*cfg.AcceptTOS {
 			log.Fatal("acceptTOS must be set to true in the config file")
 		}
 		p, err = proxy.New(cfg, []byte(*passphraseFlag))